rename: GameService 패키지 service로 이동
